[DEV] Fix main menu locations buttons, update hell location, refactor SingleMatch

+++ Assets/Scripts/TonPlay.Client.Roguelike/Network/Response/GameSessionPostBody.cs
@@ -5,6 +5,6 @@ namespace TonPlay.Client.Roguelike.Network.Response
     [System.Serializable]
     public class GameSessionPostBody
     {
-        public int surviveMills;
+        public long surviveMills;
     }
 }
